refactor: Created PasswordForm to be shared with other pages
