+++ src/client/components/PasswordForm.stories.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Meta } from '@storybook/react';
+
+import { PasswordForm } from './PasswordForm';
+
+export default {
+  title: 'Components/PasswordForm',
+  component: PasswordForm,
+} as Meta;
+
+export const Default = () => <PasswordForm submitUrl="" fieldErrors={[]} />;
+Default.storyName = 'default';
